api-server: http: order-book: Add relayer fees handler
