+++ workers/api-server/src/http.rs
@@ -15,13 +15,7 @@ use admin::{
     AdminWalletMatchableOrderIdsHandler, IsLeaderHandler,
 };
 use async_trait::async_trait;
-use common::types::{
-    MatchingPoolName,
-    gossip::{ClusterId, WrappedPeerId},
-    hmac::HmacKey,
-    tasks::TaskIdentifier,
-    token::Token,
-};
+use common::types::hmac::HmacKey;
 use external_api::{
     EmptyRequestResponse,
     http::{
@@ -40,7 +34,7 @@ use external_api::{
         network::{GET_CLUSTER_INFO_ROUTE, GET_NETWORK_TOPOLOGY_ROUTE, GET_PEER_INFO_ROUTE},
         order_book::{
             GET_DEPTH_BY_MINT_ROUTE, GET_DEPTH_FOR_ALL_PAIRS_ROUTE, GET_EXTERNAL_MATCH_FEE_ROUTE,
-            GET_NETWORK_ORDER_BY_ID_ROUTE, GET_NETWORK_ORDERS_ROUTE,
+            GET_NETWORK_ORDER_BY_ID_ROUTE, GET_NETWORK_ORDERS_ROUTE, GET_RELAYER_FEES_ROUTE,
         },
         price_report::{GET_SUPPORTED_TOKENS_ROUTE, GET_TOKEN_PRICES_ROUTE, PRICE_REPORT_ROUTE},
         task::{GET_TASK_QUEUE_PAUSED_ROUTE, GET_TASK_QUEUE_ROUTE, GET_TASK_STATUS_ROUTE},
@@ -63,8 +57,6 @@ use hyper::{
     server::conn::http1::Builder as Http1Builder, service::service_fn,
 };
 use hyper_util::rt::{TokioIo, TokioTimer};
-use num_bigint::BigUint;
-use num_traits::Num;
 use order_book::{GetDepthByMintHandler, GetExternalMatchFeesHandler};
 use price_report::{GetSupportedTokensHandler, TokenPricesHandler};
 use rate_limit::WalletTaskRateLimiter;
@@ -73,11 +65,9 @@ use std::{net::SocketAddr, sync::Arc};
 use task::GetTaskQueuePausedHandler;
 use tokio::net::{TcpListener, TcpStream};
 use util::get_current_time_millis;
-use uuid::Uuid;
 
 use crate::{
-    error::{bad_request, not_found},
-    http::order_book::GetDepthForAllPairsHandler,
+    http::order_book::{GetDepthForAllPairsHandler, GetRelayerFeesHandler},
     router::QueryParams,
 };
 
@@ -108,134 +98,6 @@ use super::{
 
 /// Health check
 pub const PING_ROUTE: &str = "/v0/ping";
-
-// ------------------
-// | Error Messages |
-// ------------------
-
-/// Error message displayed when a mint cannot be parsed from URL
-const ERR_MINT_PARSE: &str = "could not parse mint";
-/// Error message displayed when a given order ID is not parsable
-const ERR_ORDER_ID_PARSE: &str = "could not parse order id";
-/// Error message displayed when a given wallet ID is not parsable
-const ERR_WALLET_ID_PARSE: &str = "could not parse wallet id";
-/// Error message displayed when a given cluster ID is not parsable
-const ERR_CLUSTER_ID_PARSE: &str = "could not parse cluster id";
-/// Error message displayed when a given peer ID is not parsable
-const ERR_PEER_ID_PARSE: &str = "could not parse peer id";
-/// Error message displayed when parsing a task ID from URL fails
-const ERR_TASK_ID_PARSE: &str = "could not parse task id";
-/// Error message displayed when parsing a matching pool name from URL fails
-const ERR_MATCHING_POOL_PARSE: &str = "could not parse matching pool name";
-/// Error message displayed when an invalid token is parsed from a URL param
-const ERR_INVALID_TOKEN_PARSE: &str = "invalid token";
-
-// ----------------
-// | URL Captures |
-// ----------------
-
-/// The :mint param in a URL
-const MINT_URL_PARAM: &str = "mint";
-/// The :wallet_id param in a URL
-pub(super) const WALLET_ID_URL_PARAM: &str = "wallet_id";
-/// The :order_id param in a URL
-const ORDER_ID_URL_PARAM: &str = "order_id";
-/// The :cluster_id param in a URL
-const CLUSTER_ID_URL_PARAM: &str = "cluster_id";
-/// The :peer_id param in a URL
-const PEER_ID_URL_PARAM: &str = "peer_id";
-/// The :task_id param in a URL
-const TASK_ID_URL_PARAM: &str = "task_id";
-/// The :matching_pool param in a URL / query string
-const MATCHING_POOL_PARAM: &str = "matching_pool";
-
-/// A helper to parse out a mint from a URL param
-pub(super) fn parse_mint_from_params(params: &UrlParams) -> Result<BigUint, ApiServerError> {
-    // Try to parse as a hex string, then fall back to decimal
-    let mint_str = params.get(MINT_URL_PARAM).ok_or_else(|| not_found(ERR_MINT_PARSE))?;
-    let stripped_param = mint_str.strip_prefix("0x").unwrap_or(mint_str);
-    if let Ok(mint) = BigUint::from_str_radix(stripped_param, 16 /* radix */) {
-        return Ok(mint);
-    }
-
-    params.get(MINT_URL_PARAM).unwrap().parse().map_err(|_| bad_request(ERR_MINT_PARSE))
-}
-
-/// A helper to parse a token (":mint") from a URL param
-pub(super) fn parse_token_from_params(params: &UrlParams) -> Result<Token, ApiServerError> {
-    let mint_str = params.get(MINT_URL_PARAM).ok_or_else(|| not_found(ERR_MINT_PARSE))?;
-    let token = Token::from_addr(mint_str);
-    if !token.is_named() {
-        return Err(bad_request(ERR_INVALID_TOKEN_PARSE));
-    }
-
-    Ok(token)
-}
-
-/// A helper to parse out a wallet ID from a URL param
-pub(super) fn parse_wallet_id_from_params(params: &UrlParams) -> Result<Uuid, ApiServerError> {
-    params
-        .get(WALLET_ID_URL_PARAM)
-        .ok_or_else(|| bad_request(ERR_WALLET_ID_PARSE))?
-        .parse()
-        .map_err(|_| bad_request(ERR_WALLET_ID_PARSE))
-}
-
-/// A helper to parse out an order ID from a URL param
-pub(super) fn parse_order_id_from_params(params: &UrlParams) -> Result<Uuid, ApiServerError> {
-    params
-        .get(ORDER_ID_URL_PARAM)
-        .ok_or_else(|| bad_request(ERR_ORDER_ID_PARSE))?
-        .parse()
-        .map_err(|_| bad_request(ERR_ORDER_ID_PARSE))
-}
-
-/// A helper to parse out a cluster ID from a URL param
-pub(super) fn parse_cluster_id_from_params(
-    params: &UrlParams,
-) -> Result<ClusterId, ApiServerError> {
-    params
-        .get(CLUSTER_ID_URL_PARAM)
-        .ok_or_else(|| bad_request(ERR_CLUSTER_ID_PARSE))?
-        .parse()
-        .map_err(|_| bad_request(ERR_CLUSTER_ID_PARSE))
-}
-
-/// A helper to parse out a peer ID from a URL param
-pub(super) fn parse_peer_id_from_params(
-    params: &UrlParams,
-) -> Result<WrappedPeerId, ApiServerError> {
-    params
-        .get(PEER_ID_URL_PARAM)
-        .ok_or_else(|| bad_request(ERR_PEER_ID_PARSE))?
-        .parse()
-        .map_err(|_| bad_request(ERR_PEER_ID_PARSE))
-}
-
-/// A helper to parse out a task ID from a URL param
-pub(super) fn parse_task_id_from_params(
-    params: &UrlParams,
-) -> Result<TaskIdentifier, ApiServerError> {
-    params
-        .get(TASK_ID_URL_PARAM)
-        .ok_or_else(|| bad_request(ERR_TASK_ID_PARSE))?
-        .parse()
-        .map_err(|_| bad_request(ERR_TASK_ID_PARSE))
-}
-
-/// A helper to parse out a matching pool name from a URL param
-pub(super) fn parse_matching_pool_from_url_params(
-    params: &UrlParams,
-) -> Result<MatchingPoolName, ApiServerError> {
-    params.get(MATCHING_POOL_PARAM).ok_or_else(|| bad_request(ERR_MATCHING_POOL_PARSE)).cloned()
-}
-
-/// A helper to parse out a matching pool name from a query string
-pub(super) fn parse_matching_pool_from_query_params(
-    params: &QueryParams,
-) -> Option<MatchingPoolName> {
-    params.get(MATCHING_POOL_PARAM).cloned()
-}
 
 /// A wrapper around the router and task management operations that
 /// the worker may delegate to
@@ -478,6 +340,13 @@ impl HttpServer {
         );
 
         // --- Orderbook Routes --- //
+
+        // The "/order_book/relayer-fees" route
+        router.add_unauthenticated_route(
+            &Method::GET,
+            GET_RELAYER_FEES_ROUTE.to_string(),
+            GetRelayerFeesHandler::new(state.clone()),
+        );
 
         // The "/order_book/orders" route
         router.add_unauthenticated_route(
