+++ workers/api-server/src/http/wallet.rs
@@ -44,13 +44,13 @@ use util::{
 use crate::{
     compliance::ComplianceServerClient,
     error::{ApiServerError, bad_request, internal_error, not_found},
+    param_parsing::{
+        parse_mint_from_params, parse_order_id_from_params, parse_wallet_id_from_params,
+    },
     router::{ERR_WALLET_NOT_FOUND, QueryParams, TypedHandler, UrlParams},
 };
 
-use super::{
-    parse_mint_from_params, parse_order_id_from_params, parse_wallet_id_from_params,
-    rate_limit::WalletTaskRateLimiter,
-};
+use super::rate_limit::WalletTaskRateLimiter;
 
 // -----------
 // | Helpers |
