+++ workers/api-server/src/websocket/order_status.rs
@@ -8,7 +8,7 @@ use system_bus::{SystemBus, TopicReader};
 use crate::{
     auth::AuthType,
     error::{ApiServerError, not_found},
-    http::parse_wallet_id_from_params,
+    param_parsing::parse_wallet_id_from_params,
     router::{ERR_WALLET_NOT_FOUND, UrlParams},
 };
