+++ workers/api-server/src/websocket/task.rs
@@ -12,7 +12,7 @@ use system_bus::{SystemBus, TopicReader};
 use crate::{
     auth::AuthType,
     error::{ApiServerError, not_found},
-    http::{parse_task_id_from_params, parse_wallet_id_from_params},
+    param_parsing::{parse_task_id_from_params, parse_wallet_id_from_params},
     router::UrlParams,
 };
